% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{add_user_tab}
\alias{add_user_tab}
\title{Add tab for specific user role}
\usage{
add_user_tab(session, tab_id, user_role)
}
\arguments{
\item{session}{Shiny session object}

\item{tab_id}{Tab identifier}

\item{user_role}{User role}
}
\description{
Add tab for specific user role
}
