% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{create_emergency_tab}
\alias{create_emergency_tab}
\title{Create emergency procedures tab}
\usage{
create_emergency_tab()
}
\value{
UI content for emergency tab
}
\description{
Create emergency procedures tab
}
