% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_rand.R
\name{decrypt_assignment_mask}
\alias{decrypt_assignment_mask}
\title{Decrypt Assignment Mask}
\usage{
decrypt_assignment_mask(encrypted_mask, encryption_key)
}
\arguments{
\item{encrypted_mask}{Encrypted assignment mask}

\item{encryption_key}{Raw vector containing encryption key}
}
\value{
Data frame with decrypted arm labels and IDs
}
\description{
Decrypts the assignment mask to reveal arm labels
}
