% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{ehs_randomization_app}
\alias{ehs_randomization_app}
\title{EHS Pilot Study Randomization Application}
\usage{
ehs_randomization_app(
  dbname = "ehs_pilot",
  host = "localhost",
  port = 5432,
  user = "postgres",
  password = "password"
)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Database host}

\item{port}{Database port}

\item{user}{Database user}

\item{password}{Database password}
}
\description{
Complete Shiny application for managing blinded randomization in the EHS pilot study
}
