% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_rand.R
\name{generate_encrypted_arms}
\alias{generate_encrypted_arms}
\title{Generate Encrypted Assignment Strings}
\usage{
generate_encrypted_arms(arm_labels, encryption_key = NULL)
}
\arguments{
\item{arm_labels}{Character vector of actual arm labels}

\item{encryption_key}{Raw vector containing encryption key}
}
\value{
A data frame with arm labels and encrypted arm IDs
}
\description{
Creates encrypted arm identifiers for blinded randomization
}
