% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_rand.R
\name{generate_user_ids}
\alias{generate_user_ids}
\title{Generate User IDs}
\usage{
generate_user_ids(n_participants, prefix = "EHS", id_length = 6)
}
\arguments{
\item{n_participants}{Number of participants}

\item{prefix}{Character prefix for IDs (default: "EHS")}

\item{id_length}{Length of random component (default: 6)}
}
\value{
Character vector of unique participant IDs
}
\description{
Creates unique participant IDs for the study
}
