% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{get_access_level}
\alias{get_access_level}
\title{Get access level description for user role}
\usage{
get_access_level(role)
}
\arguments{
\item{role}{User role string}
}
\value{
Character description of access level
}
\description{
Get access level description for user role
}
