% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{get_user_tabs}
\alias{get_user_tabs}
\title{Get tabs available for user role}
\usage{
get_user_tabs(role)
}
\arguments{
\item{role}{User role string}
}
\value{
Character vector of available tab IDs
}
\description{
Get tabs available for user role
}
