% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_rand.R
\name{randomization_quick_start}
\alias{randomization_quick_start}
\title{Quick Start Guide for EHS Randomization System}
\usage{
randomization_quick_start()
}
\description{
This function provides a quick start guide for setting up and using
the EHS randomization system.
}
