% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymod_rand.R
\name{randomizationServer}
\alias{randomizationServer}
\title{Server Function for Randomization Module}
\usage{
randomizationServer(id, postgres_module, auth_info)
}
\arguments{
\item{id}{A character string that matches the ID used in randomizationUI}

\item{postgres_module}{A postgresModule instance}

\item{auth_info}{Reactive containing authentication information}
}
\description{
Handles the server-side logic for the randomization system
}
