% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymod_rand.R
\name{randomizationUI}
\alias{randomizationUI}
\title{UI Function for Randomization Module}
\usage{
randomizationUI(id)
}
\arguments{
\item{id}{A character string that defines the namespace for the module}
}
\value{
A list containing UI elements for different user roles
}
\description{
Creates the user interface for the randomization system
}
