% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema_rand.R
\name{retrieve_encrypted_key}
\alias{retrieve_encrypted_key}
\title{Retrieve Encrypted Key}
\usage{
retrieve_encrypted_key(
  postgres_module,
  study_id,
  key_name,
  master_password,
  user_id
)
}
\arguments{
\item{postgres_module}{A postgresModule instance}

\item{study_id}{Character string identifying the study}

\item{key_name}{Name/identifier for the key}

\item{master_password}{Character string for key decryption}

\item{user_id}{User retrieving the key}
}
\value{
Raw vector containing the decrypted key
}
\description{
Retrieves and decrypts encryption keys from the database
}
