% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehs_rand.R
\name{setup_demo_data}
\alias{setup_demo_data}
\title{Example function to setup demo data}
\usage{
setup_demo_data(postgres_module)
}
\arguments{
\item{postgres_module}{A postgresModule instance}
}
\description{
Creates demo users and initial data for testing the system
}
