% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema_rand.R
\name{setup_randomization_database}
\alias{setup_randomization_database}
\title{Setup Randomization Database Schema}
\usage{
setup_randomization_database(postgres_module, study_id = "EHS_PILOT")
}
\arguments{
\item{postgres_module}{A postgresModule instance}

\item{study_id}{Character string identifying the study}
}
\value{
TRUE if successful
}
\description{
Creates all necessary tables for the blinded randomization system
}
