% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_schema_rand.R
\name{store_encrypted_key}
\alias{store_encrypted_key}
\title{Store Encrypted Key}
\usage{
store_encrypted_key(
  postgres_module,
  study_id,
  key_name,
  encryption_key,
  master_password,
  user_id
)
}
\arguments{
\item{postgres_module}{A postgresModule instance}

\item{study_id}{Character string identifying the study}

\item{key_name}{Name/identifier for the key}

\item{encryption_key}{Raw vector containing the key}

\item{master_password}{Character string for key encryption}

\item{user_id}{User storing the key}
}
\description{
Securely stores encryption keys in the database
}
