% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymod_rand.R
\name{store_randomization_data}
\alias{store_randomization_data}
\title{Helper function to store randomization data in database}
\usage{
store_randomization_data(postgres_module, study_id, randomization_package)
}
\arguments{
\item{postgres_module}{A postgresModule instance}

\item{study_id}{Character string identifying the study}

\item{randomization_package}{List containing randomization data}
}
\description{
Helper function to store randomization data in database
}
